#175 Refactor out utility class
